requirements.txt
azure-mgmt-compute==31.*
azure-mgmt-network==25.*
azure-mgmt-resource==23.*
azure-identity==1.*
aiohttp==3.*
python-dotenv==1.0.*
